refactor: move stellar and general test routes to new 'test_routes' folder
